requirements.txt
Django~=3.1.3
line-bot-sdk~=1.17.0
django-heroku~=0.3.1
gunicorn~=20.0.4
